+++ RELEASE_NOTES.md
@@ -18,9 +18,13 @@ This version adds tap behavior to the margin next to character inputs.
 
 * `KeyboardAction` has a new `characterMargin` action.
 
+### 💡 Behavior changes
+
+* `iPadKeyboardLayoutProvider` and `iPhoneKeyboardLayoutProvider` now only adjust input sets with three rows.
+
 ### 🗑 Deprecations
 
-* `iPhoneKeyboardLayoutProvider` has new `lowerLeadingActions` and `lowerTrailingActions` that take the row as parameter to resolve the character margin.
+* `SystemKeyboardLayoutProvider` has new margin action functions.
 * `SystemKeyboardLayoutProvider` has been redesigned to let all functions that take multiple parameters with an initial context, place the context parameter last. This gives a cleaner and more harmonized public api.   
 * `SystemKeyboardLayoutProvider` `inputs(for:)` has been renamed to `inputRows(for:)`.